+++ src/ui/react/weg/modules/shared/types.ts
@@ -5,6 +5,7 @@ export type SeparatorWegItem = Extract<WegItem, { type: "Separator" }>;
 export type MediaWegItem = Extract<WegItem, { type: "Media" }>;
 export type StartMenuWegItem = Extract<WegItem, { type: "StartMenu" }>;
 export type ShowDesktopWegItem = Extract<WegItem, { type: "ShowDesktop" }>;
+export type TrashBinItem = Extract<WegItem, { type: "TrashBin" }>;
 
 /** @alias */
 export type SwItem = WegItem;